+++ src/unicode/property.h
@@ -20,6 +20,7 @@
 #include "../misc/enum_util.hpp"
 #include "../misc/result.hpp"
 #include "../misc/string_ref.hpp"
+#include "emoji_seq.hpp"
 #include "set_builder.h"
 
 #include "ucp_general_category_def.in"
@@ -117,6 +118,12 @@ inline CodePointSet getPropertySet(const Property property) {
 }
 
 bool isExtendedPictographic(int codePoint);
+
+Optional<EmojiProperty> parseEmojiProperty(StringRef ref);
+
+const char *toString(EmojiProperty p);
+
+EmojiProperty getEmojiProperty(StringRef ref);
 
 } // namespace arsh::ucp